pyproject.toml
[project]
name = "xgboost-simple-gui-app"
version = "0.1.0"
description = "A simple XGBoost training app with GUI"
requires-python = ">=3.12"
dependencies = [
    "xgboost>=2.0.3",
    "scikit-learn>=1.4.0",
    "pandas>=2.2.0",
    "numpy>=1.26.3",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.3",
    "pymongo>=4.6.0",
    "ruff>=0.12.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["python_api"]

[tool.uv]
dev-dependencies = [
    "ipython>=8.20.0",
]
